# apps/user-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      timezone:
        type: string
    required:
    - email
    - first_name
    - last_name
    type: object
  models.Household:
    properties:
      address:
        type: string
      coordinates:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
      wifi_ssid:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      phone:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8084
info:
  contact: {}
  description: Микросервис для управления пользователями и домохозяйствами
  title: User Service API
  version: "1.0"
paths:
  /households:
    post:
      consumes:
      - application/json
      description: Создает новое домохозяйство для пользователя
      parameters:
      - description: Данные домохозяйства
        in: body
        name: household
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Household'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать домохозяйство
      tags:
      - Households
  /households/{id}:
    get:
      description: Возвращает данные домохозяйства по его ID
      parameters:
      - description: ID домохозяйства
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Household'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить домохозяйство по ID
      tags:
      - Households
  /users:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать пользователя
      tags:
      - Users
  /users/{id}:
    get:
      description: Возвращает данные пользователя по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить пользователя по ID
      tags:
      - Users
swagger: "2.0"